requirements.txt
Flask==2.0.3
tensorflow==2.16.1
torch==2.2.0
opacus==1.5.2
python-dotenv==0.19.2
joblib==1.1.0
numpy==1.24.3
pandas==1.5.3
tensorflow-addons==0.23.0
packaging==24.1
typeguard==2.13.3
wheel==0.44.0
Werkzeug==2.0.3
google-generativeai